requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.2=py39he3068b8_1
ca-certificates=2022.2.1=hecd8cb5_0
certifi=2021.10.8=py39hecd8cb5_2
freetype=2.11.0=hd8bbffd_0
giflib=5.2.1=haf1e3a3_0
intel-openmp=2021.4.0=hecd8cb5_3538
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
ncurses=6.3=hca72f7f_2
numexpr=2.8.1=py39h2e5f0a9_0
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39he9d5cce_0
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.9.7=h88f2d9e_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
readline=8.1.2=hca72f7f_1
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=h707629a_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h7bc2e8c_0
tzdata=2021e=hda174b7_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
